05-01-2026-1/requirements.txt
# requirements.txt
fastapi
uvicorn
playwright
pillow
imageio
numpy
python-multipart
jinja2
sqlalchemy
databases[aiosqlite]
python-jose[cryptography]
passlib[bcrypt]
httpx
phonenumbers
argon2-cffi